requirements.txt
Django==1.11.10
djangorestframework==3.7.7
celery==3.1.26.post2
django-celery==3.3.0
yunpian-python-sdk==1.0.0
django-cors-headers==3.0.1
django-redis==4.10.0